feat: Add support for PyTorch and FashionMNIST dataset
The changes in this commit introduce support for PyTorch and the FashionMNIST dataset in the ML_basics.ipynb notebook. The main changes are:

1. Added code to download and load the FashionMNIST dataset using PyTorch's datasets module.
2. Created data loaders for the training and test data using PyTorch's DataLoader class.
3. Defined a neural network model using PyTorch's nn.Module class, which includes a flattening layer and a stack of linear and ReLU layers.
4. Specified the loss function (nn.CrossEntropyLoss) and optimizer (torch.optim.SGD) to be used for training the model.

These changes provide a foundation for working with PyTorch and the FashionMNIST dataset in the notebook, which can be used for further exploration and development of machine learning models.